+++ inc/header.php
@@ -28,7 +28,10 @@ header("Content-type: text/html;charset=UTF-8");
         <script src="lib/datepicker/js/bootstrap-datepicker.js" ></script>
         <script src="lib/highcharts/js/highcharts.js" ></script>
         <script src="lib/highcharts/js/modules/exporting.js" ></script>
-        <script>var current_version = "<?php echo VERSION; ?>";</script>
+        <script>
+            var current_version = "<?php echo VERSION; ?>";
+            var latest_version = "<?php echo VERSION; ?>"; // will be overiden by latest.js
+        </script>
         <script src="js/app.js" ></script>
         <script src="http://serphacker.com/serposcope/latest.js" ></script>